feat: api 요청 성공

+++ Retrofit_Error_Handling/app/src/main/java/com/example/retrofit_error_handling/presentation/model/BookUiModel.kt
@@ -5,5 +5,5 @@ data class BookUiModel(
     val contents: String,
     val url: String,
     val price: Int,
-    val isbn: Int
+    val isbn: Long
 )
